requirements.txt
Django==3.2.8
PyMuPDF ==1.23.26
Pillow==10.2.0
protobuf==5.26.0
python-dotenv==1.0.1
PyPDF2==3.0.1
google-generativeai==0.3.2